Don't apply the rewrite-as-range optimization if field is multivalued (#84535)
This fixes a bug when trying to optimize date histograms into date range aggregations.  For multivalued fields with a date query, the optimization would only capture the values of the field that matched the query, rather than all values of the document as a non-rewritten date histogram would do.  This PR disables the query-level optimization when there are multi-valued fields in the index, as there's no good way work around it. Index level range rewrite optimizations are unaffected.

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>